Moved class CLogOut to project 3d fixes TweetWallFX/TweetwallFX#171
